Better support for update strategy RollingUpgrade
An issue was opened when a tester tried to add a license secret after
the cluster was up.  Adding the secret causes the statefulset controller
to apply the update in a rolling update fashion.  This didn't work so
well, the update eventually froze while it was trying to restart
vertica.  The changes here are meant to address that.

- change the readiness probe back to using the 'select 1' canary query.
This was critical in making the rolling update strategy to behave better.
The old readiness probe was causing us to go onto the next pod to soon.
If you only have 3 pods in a k-safety 1 system, this would cause us to
lose two pods, which loses quorum so the 3rd pod goes down to.
- the 'select 1' canary query was used before but was changed when we
added support for the superuser password.  We didn't want to hardcode
the password.  We overted that here by getting the password from a file
in the container; the file is /etc/podinfo/superuser-passwd.  It is
included in the projected volume that currently contains things from the
downward API.
- use 'OnDelete' for the UpdateStrategy for k-safety 0.  Any pod loss
will take down vertica on the remaining pods.  So supporting a rolling
update strategy seemed pointless.
- add a new option 'restartTimeout' that controls what timeout to use
with admintools when restarting pods
- during restart we will kill any outstanding vertica process ahead of
doing the actual restart.  This seemed to help the restart process, but
is mitigated by the readiness probe change.  I kept it in here because I
think it is still good to have.