Save contents of /opt/vertica/config/licensing
When running the vertica-k8s container all of the contents in
/opt/vertica/config are lost on initial startup because we use a PV to
back that.  There are certain files that we want to keep and must be
copied back over when the container starts.  We already stored the
contents of /opt/vertica/config/licensing in /home/dbadmin/licensing/ce,
we are going to copy that directory over each time the container starts.