Triggers e2e tests on push, PR comment and manually (#15)
A new CI file has been added. When we push on the main branch or write a specific
comment on the PR it triggers e2e tests. You can also triggers them manually by
going on the section "Actions" on github. 

Therer were some problems with create bucket as aws was trying to retrieve the
region and in an isolated env as github it was causing an error so we set 
AWS_EC2_METADATA_DISABLED to true for aws to ignore that. 

We remove the agent because it was preventing us to scale out.