Add support to run generic sidecars
This adds the ability run a custom sidecar alongside the vertica server.
A new property was added to the CR to allow this:  spec.sidescars[]

When the operator adds the new container, it will automatically include
the volume mounts that were added to the server container.  This was it
automatically has access to the same PV.

A new container was added for testing purposes.  Called the vlogger, it
will send output from vertica.log to stdout in a sidecar.  A few e2e
tests use this to ensure we can run with a sidecar.  This container is
not intended to be externalized in Vertica's dockerhub.

To allow this new image to be used in the e2e tests, the kustomize setup
was changed to allow for this change to be set and overwritten.  Some of
the VerticaDB specs that we use for tests have a new
`kustomize-vlogger-image` placeholder for this purpose.