Support upgrade when cluster is in bad state (#19)
Previously a call to stop_db in upgrade_vertica returned an error when the
previous vertica image was a bad one or when the cluster was down. So now we
first check if all pods are running (because if a bad image is running inside
them their status will obviously not be "Running"). if they are not, we skip
the call to stop_db and proceed to update the bad image with the new one. if
the pods are running, we check if there is at least one node of the cluster UP.
if there is not we skip the stop_db call.

Moreover these 2 scenarios have been added to upgrade-vertica test.