Allow e2e parallelism to be overridden (#20)
Now we can pass a parameter to override the parallel field in kuttl. Run the
workflow manually with workflow_dispatch event and specify the number of tests
you want to run at the same time. if you do not specify any number the e2e
tests will use parameters set in kuttl-test.yml.