Reduce disk usage of e2e tests (#33)
One of the e2e tests was failing consistently because we were
hitting disk full. I reduced the disk usage by changing the kind
cluster we set from 2 nodes to 1 node. This helps reduced the
amount of disk needed because we only had to push the containers
to 1 k8s node now instead of 2.

Also changed the name of the tests in GitHub actions from build
to e2e/ut.