e2e test changes to work in public cloud (#35)
A series of changes to the e2e tests so that they can run in public cloud (i.e.
GKE, AKS, etc.). The bulk of the changes was to change the test suite so there
was a single minio tenant for all testcases. Previously, each testcase would
create their own tenant. But this caused a large amount of disk space to be
consumed, causing disk full errors on GKE and AKS.

The convention is that each testcase must create their own bucket in the shared
minio tenant. The bucket name should be the name of the testcase to avoid any
collision.