Integrate webhook into the operator pod (#43)
This integrates the webhook into the operator pod. We previously had them
separate because the webhook was cluster scoped and the operator was namespace
scoped. To solve that issue, when we deploy the operator we will create
ValidatingWebhookConfigurations and MutatingWebhookConfigurations in the
following way:

- prefixed with the namespace name to provide uniqueness
- use a namespaceSelector, so that the config only applies to the operator in a
  given namespace

In order to use the namespaceSelector, you need to use the
NamespaceDefaultLabelName feature gate (enabled by default on k8s 1.21.1) or
ensure the namespace has the following label 'kubernetes.io/metadata.name=name'.

The e2e tests now run on k8s 1.21.1 to take advantage of this feature gate.

Some additional changes:
- removed the webhook container information
- removed the webhook helm chart
- removed the webhook executable
- 'make manifests' no longer generates the webhook information. This allowed us
  to add the namespaceSelector to webhook.