Fix to allow create DB to work on AWS s3 (#47)
The operator had set AWS timeouts low so that during create_db we get quick
feedback in case the s3 endpoint isn't setup correctly. These low timeouts were
interfering with the create_db in such a way as preventing a successful create
on aws s3. The timeouts were removed in this commit. This also means we cannot
reliable check for this error in the createdb-failures testcase.

A few other minor changed in included:

- adding event recording for when the s3 auth parms copy fails. It is possible
  that if the secret was not setup correctly that the copy would contain
  undisplayable characters causing the entire copy to fail. We log an event now
  so that end users have a better idea of what issue the operator is hitting.
- when running the operator with the 'make run' target, we need the ability to
  disable the webhook. Otherwise, it will fail to start due to lack of TLS
  certs causing the operator to fail. Added a check for the ENABLE_WEBHOOKS
  environment variable.