Relax cert manager dependency (#46)
This adds to the helm chart an option to allow specifying your own certs for
use with the webhook. The default behaviour is to continue to rely on
cert-manager. But if the webhook.tlsSecret is specified then cert-manager isn't
used.

To support this, changes had to be made to the generation of the helm-chart.
- we no longer keep all manifests in a single file. We use a tool called
kuberenetes-split-yaml that splits up a giant yaml into individual files -- one
for each Kind. Splitting this allowed for easier manipulation of the manifests since
we template out the cert-manager ones if webhook.tlsSecrets and add a caBundle
to the webhook configs.
- logic for creation of the helm chart was getting too big for the Makefile, so
was moved to its own script (create-helm-charts.sh)

A new e2e test was added that deploys the webhook with custom certs. The certs
are still generated by cert-manager, but it is done outside the helm chart
install so mimics what a customer would do if they were using their own certs.