Add ability to mount custom certs (#54)
This provides the ability to mount any number of custom certs in the Vertica
container. Each cert will be mounted in a well defined location within the
container (/certs/<certName>/<key>). This will be available with a new
parameter called spec.certSecrets. A list of Secret names can be specified.

Here is an example of a CR that takes advantage of that:

apiVersion: vertica.com/v1beta1
kind: VerticaDB
metadata:
  name: verticadb-sample
spec:
  communal: {…}
  certSecrets:
    - name: mtls
    - name: aws-cert

It specifies two secrets: mtls and aws-cert. The keys within these Secrets will
be mounted at /certs/mtls and /certs/aws-cret respectively. If the keys of the
secret change, the mount points will be automatically updated to reflect the
new value without having to restart the pod.