Handle automatic upgrades (#51)
This adds the ability to automate an upgrade. In the v11.0 release, we
published a script to use to upgrade a VerticaDB CR. This change pushes much of
the logic of that script into the operator.

If you want to upgrade vertica, you simply update the .spec.image field in the
VerticaDB CR. The operator will detect when this happens and go into a mode
that handles the upgrade.

New events are logged to know when the operator is doing this:

$ kubectl describe vdb sample
 
...<snip>...
Events:   
  Type    Reason                   Age    From                Message 
  ----    ------                   ----   ----                ------- 
  Normal  ImageChangeStart         5m10s  verticadb-operator  Vertica server image change has been initiated to 'vertica-k8s:11.0.1-0'
  Normal  ClusterShutdownStarted   5m12s  verticadb-operator  Calling 'admintools -t stop_db' 
  Normal  ClusterShutdownSucceeded 4m08s  verticadb-operator  Successfully called 'admintools -t stop_db' and it took 56.22132s
  Normal  ClusterRestartStarted    4m25s  verticadb-operator  Calling 'admintools -t start_db' to restart the cluster
  Normal  ClusterRestartSucceeded  25s    verticadb-operator  Successfully called 'admintools -t start_db' and it took 240s
  Normal  ImageChangeSucceeded     5s     verticadb-operator  Vertica server image change has completed successfully.

A new status condition is added so that you can programmatically wait for the
upgrade to complete, or timeout and rollback the upgrade.

To wait for the operator to get into the upgrade mode, you can issue this:
kubectl wait --for=condition=ImageChangeInProgress=True vdb/v --timeout=x

To wait for the operator to finish the upgrade, you can issue this:
kubectl wait --for=condition=ImageChangeInProgress=False vdb/v --timeout=x

where v is the name of the CR and x is a timeout value.

A new reconciler was added to handle the upgrade (called
ImageChangeReconciler). It uses parts of ObjReconciler and RestartReconciler to
avoid code duplication of actions those reconcilers already handle.

A few other minor changes to mention:    
- removed the AutoRestartVertica column when we issue 'kubectl get vdb'. It
  didn't make sense to report that state now that it isn't as important with  auto upgrade
- include vertica version, pulled from the rpm, in the vertica container. This
  can be used in the future if we want to add logic about what following the
  proper upgrade path.