Allow custom endpoints and credentials in e2e tests (#55)
Refactor e2e tests to allow for custom endpoints and credentials. This is prep
work for issue #45. We want to split the e2e tests so that half of them access
minio through https://. This adds another endpoint and updates half of the
testcases to use that one. For now both endpoints are the same.

Custom endpoints and credentials can now be used with the e2e tests. They are
provided through a config file that create-kustomize-overlay.sh uses. The setup
of the communal endpoint (minio by default) is separated from kuttl, and is now
a make dependency when running the e2e tests with make. More details were added
to DEVELOPER.md.