Allow region to be specified (#61)
This adds a new parameter to .spec.communal to allow the s3 region to be
specified. Prior to this, we would always use the Vertica default of us-east-1.

Here is a sample CR that specifies us-west-2:

apiVersion: vertica.com/v1beta1
kind: VerticaDB
metadata:
  name: sample
spec:
  communal:
    path: "s3://vertica-fleeting-us-west-2/mspilchen"
    endpoint: "https://s3.us-west-2.amazonaws.com"
    credentialSecret: s3-auth
    region: us-west-2
  kSafety: "0"
  subclusters:
    - name: sc1
      size: 3

The webhook was updated to fill in the region if it wasn't set. This way it
will be clear what the default region rather than relying on the default set in
the server. New events were added to make it more visible when the region is
misconfigured.