Fix timing issue with install reconcile
There was a small timing window that is closed with this change.  It was
caught in the crash-before-createdb testcase.  The order in install was
to add hosts to admintools.conf then accept the EULA.  If the pod is
rescheduled between these, we would attempt a re_ip before coming back
and completing the install.  But the re_ip kept failing because the EULA
wasn't accepted.   The solution was to accept the EULA *before* adding
the hosts to the admintools.conf.