Allow communal path to be configurable in e2e tests (#66)
This is a big refactor to decouple the e2e tests from minio. The idea is that
by decoupling minio we can run any communal endpoint (e.g. AWS). This will
allow us to test the e2e tests against endpoints such as AWS, different regions
etc.

A bulk of the change was to change how the kustomize overlay files are created.
This allows us to substitute a different endpoint for the tests. The tests are
now configurable by:

- communal endpoint
- s3 bucket and communal path prefix
- access and secret keys
- whether or not a CA cert is needed to connect to the endpoint
- region