Fix image change when it includes a scale down
This fixes a bug when changing the image with a reduction in the size of
a subcluster.  Prior to this, we would reduce the sts for the subcluster
and not run db_remove_node to remove it from the Vertica cluster.
Because we didn't do this, we could not restart the cluster.

The fix was to preserve the size of the sts and just update its image.
This required that we delete all of the pods after the image change so
they can be rescheduled with the new image.