kind.sh term to delete registry by default (#79)
The -x option in kind.sh will now be used to skip killing of the registry.

A new option was added to kind.sh (-r) to allow overriding the port number of
registry.  It defaults to port 5000.

Also changed the default deployment method for e2e tests to be helm. When
running the e2e tests in the CI it will randomly pick between helm and olm.