Fix the upgrade-vertica e2e test
The upgrade-vertica testcase cycles the vertica-k8s image with tags
latest and 11.0.0-0-minimal.  Yesterday, we changed latest to point
to 11.0.1.  So the test was now doing upgrades and downgrades.  And
Vertica dies if you do a downgrade.  Changing the test to cycle
through images 11.0.0-0-minimal and 11.0.0-0.  This way we don't do
any downgrades.