Fix the upgrade-vertica e2e test (#86)
The upgrade-vertica testcase cycles the vertica-k8s image with tags
latest and 11.0.0-0-minimal. Yesterday, we changed latest to point
to 11.0.1. So the test was now doing upgrades and downgrades. And
Vertica dies if you do a downgrade. Changing the test to cycle
between 11.0.0-0-minimal and a bogus tag, this way we don't do any
downgrades. I wanted to cycle through images 11.0.0-0-minimal and
11.0.0-0 but I was hitting disk full during CI.