Communal storage support for webhdfs://, gs://, and azb:// (#87)
This expands the different communal storage formats that we support with the
operator now. It adds HDFS through the webhdfs:// scheme -- wire-encryption
(aka swebhdfs://) and kerberos is not yet supported. It also supports
additional cloud storage options -- Google Cloud Storage (gs://) and Azure Blob
Storage (azb://).

To use webhdfs:// you need to create a ConfigMap with all of the hadoop config
files -- it can be taken from /etc/hadoop. Then setup the VerticaDB with the
.spec.communal.hadoopConfig pointing to the ConfigMap that you created and
.spec.communal.path referencing the namenode.

Here is an example that assumes the ConfigMap that you created was called
hadoop-conf:

apiVersion: vertica.com/v1beta1
kind: VerticaDB
metadata:
  name: sample
spec:
  communal:
    hadoopConfig: hadoop-conf
    path: webhdfs://hdfs-ci-namenode-0.hdfs-ci-namenode.kuttl-e2e-hadoop.svc.cluster.local:50070/client-access
  subclusters:
  - name: sc1

When using Google Cloud Storage (gs://) you first need to get a HMAC key for
your Google Cloud Storage account. The HMAC key you get will be a access key
and secret key pair. Add the keys to a Secret with key names accesskey and
secretkey.

An example VerticaDB setup for Google Cloud Storage looks like this. This
assumes a secret name communal-creds exists that has the HMAC keys.

apiVersion: vertica.com/v1beta1
kind: VerticaDB
metadata:
  name: sample
spec:
  communal:
    credentialSecret: communal-creds
    path: gs://vertica-fleeting/mspilchen/db
  subclusters:
  - name: sc1

When using Azure Blob Storage (azb://) the credentials are a bit different then
AWS/GS. You first have to decide whether to connect with the account key or a
shared access signature (SAS). All of the credentials are stored in a Secret
with the following keys: accountKey, accountName, sharedAccessSignature, and
blobEndpoint. You must set accountName and one of accountKey or
sharedAccessSignature. If you use an emulator such as Azurite, you can use
blobEndpoint, prefixing with http:// if you want to access with http.

Here is example that creates a secret for Azure.

$ kubectl create secret generic communal-creds --from-literal=accountKey=key1 \
--from-literal=accountName=devopsvertica

An example VerticaDB that uses the above credentials is as follows:

apiVersion: vertica.com/v1beta1
kind: VerticaDB
metadata:
  name: sample
spec:
  communal:
    credentialSecret: communal-creds
    path: azb://devopsvertica/vertica-fleeting/mspilchen/db
  subclusters:
  - name: sc1

The CI has been expanded so that the e2e tests run on webhdfs:// and azb://. We
use a CI specific helm chart to run HDFS and use the Azurite emulator for
azb://. The e2e tests were modified quite heavily so that they can run against
any generic communal endpoint.