Support the swebhdfs:// HDFS scheme (#88)
This adds support for HDFS with wire-encryption (swebhdfs:// scheme). In order
to use this you need to include the trust store and CA bundle in a Secret so
that it is mounted inside the Vertica container.

You can put both in a Secret using the following command:

$ kubectl create generic hadoop-certs --from-file=keystore.jks --from-file=ca-bundle.pem 

The HDFS config files needs to be setup and packaged in a single ConfigMap.  If
you have working config files already, such as in the /etc/hadoop directory,
you can easily create it as follows:

$ kubectl create configmap hadoop-conf --from-file=/etc/hadoop

This will create a single ConfigMap with multiple keys, each one being one of
the config files from that directory.

The location of the truststore in ssl-client.xml (from hadoop-conf) needs to be
set to a path relative to the container. If you follow the example CR below,
the location would be:

<property>
  <name>ssl.client.truststore.location</name>
  <value>/certs/hadoop-certs/keystore.jks</value>
</property>

Create the VerticaDB CR as follows. The caFile parameter needs to be set to
point to the CA bundle that is part of the hadoop-certs Secret.  Here is an
example yaml:

apiVersion: vertica.com/v1beta1
kind: VerticaDB
metadata:
  name: sample
spec:
  communal:
    path: "swebhdfs://hdfs-namenode-0.hdfs-namenode.default.svc.cluster.local:50070/verticadb"
    hadoopConfig: hadoop-conf
    caFile: /certs/hadoop-certs/ca-bundle.pem
  certSecrets:
    - name: hadoop-certs
  subclusters:
    - name: sc1
      size: 1