Allow custom mount points to be added to vertica (#89)
This adds the ability to add custom volume mounts to the
Vertica container. A new spec.volumeMounts parameter was
added to achieve this. It works alongside the
spec.volumes parameter.

Here is an example that mounts the path /tenants in the
Vertica container using an existing PVC.

apiVersion: vertica.com/v1beta1
kind: VerticaDB
metadata:
  name: verticadb-sample
spec:
  communal:
    ...
  subclusters:
  - name: sc1
  volumeMounts:
  - mountPath: /tenants
    name: tenants
  volumes:
  - name: tenants
    persistentVolumeClaim:
      claimName: vertica-pvc

The volumeMounts parameter takes the standard set of
options for volume mounts (e.g. same as
pod.spec.containers.volumeMounts).

The mount path for the new volume mount cannot conflict
with any existing. The webhook was updated to detect any
conflicts.

To avoid a circular dependency with the imports, I had
to move a few helper functions in the paths package to
api/v1beta1.