Fix scale down timing bug (#90)
This fixes a timing scenario with scale down that can cause corruption with the
admintools.conf.

This can occur if during ‘admintools -t db_remove_node’ another change is made
to the VerticaDB CR to scale down again. For instance, if we patched the CR to
scale a subcluster from 3 to 2. The operator will invoke ‘admintools -t
db_remove_node’. We hit the issue if before admintools returns the VerticaDB is
patched again to scale the subcluster from 2 to 1.

If you hit this problem, you will see a message like this in the operator log
when it tries admintools commands:

Error in /opt/vertica/config/admintools.conf?: No option 'v_verticadb_node0005' in section: 'Nodes'

The fix for this is that the uninstall will be requeued if we detect another
scale down has occurred. This will force us to call ‘admintools -t
db_remove_node’ for the 2nd scale down before we drive uninstall logic.