Remove the Kerberos initContainer hack (#93)
The Vertica engine has very specific requirements for the Kerberos keytab: it
must have read/write permissions for dbadmin and no one else. It isn't possible
to mount a file in k8s with specific requirements. I originally added a hack
that created an initContainer and copied the keytab so that the permissions
could be set. I'm removing that here as we now have a config knob in the engine
to relax the check. We just have to set KerberosEnableKeytabPermissionCheck =
0, and the engine won't care about keytab permissions. This also means we don't
need to restart the pod whenever the keytab changes as the change will get
automatically picked up whenever the Secret changes. Since this relies on an
engine change, I put in a version check that Kerberos cannot be used except on
11.0.2.