Handle concurrent updates to rbac during e2e tests (#100)
This closes a timing scenario with e2e when creating common cluster scoped
objects, like a ClusterRole. If two tests create the objects at the same time
(via kubectl apply), it is possible to get a failure.  The solution was to
ignore any failures when creating rbac and check the results via an assert.