Image change can be marked complete before restart of the pods (#101)
This closes a timing window during upgrade where we can indicate the image
change has finished before completing the restart of the pods. There is a check
during the restart portion that calls areAllPodsRunningAndZeroInstalled. This
returns true if none of the pods exist yet, which is not correct. State that
the function uses was being set after checking if the pod exists. If the pod
doesn't exist, the state was left unset. The timing hole is closed now that the
state is being set correctly.