Allow olm install in OpenShift (#102)
setup-olm.sh and deploy-olm.sh are used to create the catalog and deploy the
operator using olm. They have been modified to handle OpenShift as it requires
the catalog to be created in the default namespace openshift-marketplace.
Moreover, OpenShift comes with olm in openshift-operator-lifecycle-manager
namespace so we made sure to not try to install the olm-operator.