Use different Vertica version for e2e upgrade tests (#106)
upgrade-vertica-ks-0/1 fail on some platforms (OpenShift, EKS) that have more
restrictive permissions for the PV.  These tests upgrade from version 11.0.0.
The docker entry for the 11.0.1 Vertica container does a chown of all of the PV
mounts.  We are going to run those tests with that version to support the
platforms we had issues with.