Fixes for kuttl when VERTICA_IMG env not set (#112)
When running kuttl without setting VERTICA_IMG, it was using
vertica/vertica-k8s:latest as the image.  There was also some config
setting to set the default image to use if one isn't set.  We will now
use the image that `make echo-images` reports if the env. variable isn't
set.