Add scaffolding for online image change (#123)
This sketches out the flow for the online image change. Online image change is
still not functional, but it lays out the structure of the code that I will
fill in on subsequent PRs.

This also does more refactoring. There were additional changes I wanted to
reuse in offlineimagechange_reconcile.go. Those were moved to
ImageChangeInitiator, and that struct was renamed to ImageChangeManager.