Online image change: route to temporary subcluster to keep online (#131)
This is the next set of changes for online image change. It will
route to a temporary subcluster, called a transient, so that client
connections connect to an up node. It will automatically route back
to the original subcluster once the subcluster is back up. It will
also process secondary subclusters that they are brought back up.

This includes some rework in the onlineimagechange_reconciler to cut
down on the amount of code duplication.

Added the ability to specify a template for the transient cluster.
This will get created when the image change starts and will get
cleanup up when the image change is done.