Online image change: allow reuse of existing sc for temporary routing (#133)
This is the next set of changes for online image change. The CR parm
TransientSubclusterTemplate was renamed to TemporarySubclusterRouting. The parm
can be used to provide a template of a subcluster to use for temporary routing
while subclusters are down. Or the parm can also be used to specify an existing
subcluster. This later option may be useful to those that want to just reuse
existing subclusters.

This PR also cleans up how we route traffic to the subclusters. We previously
had relied on a 'transient' label. But now we route to service name or
subcluster name.