Create a custom scc to run VDB on Openshift (#129)
Previously we were running PODs with privileged scc so we decided to
use a less restrictive scc like anyuid. Given that anyuid does not
support SYS_CHROOT and AUDIT_WRITE capabilities that our vdb cluster
use on OpenShift, we created a new scc called anyuid-extra which is
very similar to anyuid but support both capabilities.